Refactor: Refactoring App Navigators
