+++ src/components/HomeMainTitle.tsx
@@ -3,11 +3,10 @@ import {View, Text, StyleSheet} from 'react-native';
 import PoppinSvg from '../assets/icons/poppin.svg';
 import AlarmOffSvg from '../assets/icons/alarmOff.svg';
 import Text24B from '../styles/texts/headline/Text24B.ts';
-// @ts-ignore
 const HomeMainTitle = ({text1, text2}) => {
-  const handlePress = () => {
-    // 예: navigation.navigate('SomeScreen');
-  };
+  // const handlePress = () => {
+  //   // 예: navigation.navigate('SomeScreen');
+  // };
 
   return (
     <>
