+++ src/types/AuthNavigatorParamList.ts
@@ -1,0 +1,14 @@
+export type AuthNavigatorParamList = {
+  MainTabNavigator: undefined;
+  ProfileEdit: undefined;
+  PasswordChange: undefined;
+  FAQ: undefined;
+  KeywordAlarm: undefined;
+  UserRegister: undefined;
+  Policy: undefined;
+  ReviewWrite: undefined;
+  OperatorRegister: undefined;
+  MemberDelete: undefined;
+  PreferenceSetting: undefined;
+};
+// Path: src/types/RootStackParamList.ts
